通过riscv64编译
